Revert "改为float64"

+++ nn/layer/activation/relu.go
@@ -17,7 +17,7 @@ func NewReLU() *ReLU {
 	return &layer
 }
 
-func LoadRelu(_ consts.DeviceType, name string, _ map[string]*pb.Dense, _ map[string]float64) layer.Layer {
+func LoadRelu(_ consts.DeviceType, name string, _ map[string]*pb.Dense, _ map[string]float32) layer.Layer {
 	var layer ReLU
 	layer.base = new("relu")
 	layer.name = name
